+++ src/test/java/com/digitalserverhost/plugins/utils/SimpleTest.java
@@ -1,0 +1,15 @@
+package com.digitalserverhost.plugins.utils;
+
+import org.junit.jupiter.api.Test;
+import org.mockbukkit.mockbukkit.MockBukkit;
+
+public class SimpleTest {
+    @Test
+    void testMock() {
+        System.out.println("Starting SimpleTest");
+        MockBukkit.mock();
+        System.out.println("MockBukkit initialized");
+        MockBukkit.unmock();
+        System.out.println("MockBukkit unmocked");
+    }
+}